Made manage post scoring and manage score periods modal utilities call new rescore function
