+++ src/pieces/discussion/scoring/rescorePeriods.ts
@@ -6,6 +6,7 @@ import { getChannelInMainGuild } from "../../../generalUtilities/getChannelInMai
 import { wait } from "../../../generalUtilities/wait";
 import { addCommentScoreToPeriod, addCommentToPosterScore, addPostScoreToPeriod, findMessagePeriod } from "./scoreActionUtilities";
 import { MessageScoreData, scoreDiscussionContent } from "./scoringUtilities";
+import { addScorePeriodArrays } from "../../../generalUtilities/ScorePeriodUtilities";
 
 export async function rescoreDiscussion(client: Client, forumId: string, discussionSpecs: DiscussionSpecs, staffId: string) {
 
@@ -54,7 +55,7 @@ function wipeStudentScores(period: ScorePeriod) {
     period.studentScores = new Map<string, StudentScoreData>();
 }
 
-export async function scoreThread(client: Client, threadId: string, discussionSpecs: DiscussionSpecs, staffId: string): Promise<ScorePeriod[]> {
+async function scoreThread(client: Client, threadId: string, discussionSpecs: DiscussionSpecs, staffId: string): Promise<ScorePeriod[]> {
     
     const clonedSpecs = loadDash.cloneDeep(discussionSpecs) // Make a deep copy so we dont overwrite when scoring multiple threads
 
@@ -139,77 +140,6 @@ function removeMessagesAfterDate(messages: Message[], date: Date) {
 function removeMessagesBeforeDate(messages: Message[], date: Date) {
     while(messages.length && messages[0].createdAt.valueOf() > date.valueOf()) {
         messages.shift()
-    }
-}
-
-// Adding Score Period Arrays
-function addScorePeriodArrays(scorePeriodsA: ScorePeriod[], scorePeriodsB: ScorePeriod[]) {
-
-    if(scorePeriodsA.length !== scorePeriodsB.length) {
-        throw new Error("ERROR CANNOT ADD TWO SCORE PERIOD ARRAYS OF DIFFERENT LENGTH");
-    }
-
-    const combinedScorePeriods = scorePeriodsA.map((_scorePeriod, index) => {
-        return addScorePeriods(scorePeriodsA[index], scorePeriodsB[index])
-    })
-
-    return combinedScorePeriods;
-}
-
-export function addScorePeriods(scorePeriodA: ScorePeriod, scorePeriodB: ScorePeriod) {
-    
-    const EMPTY_SCORE_VALUE: StudentScoreData = {
-        score: 0,
-        numPosts: 0,
-        numIncomPost: 0,
-        numComments: 0,
-        numIncomComment: 0,
-        awardsRecieved: 0,
-        penaltiesRecieved: 0
-    }
-
-    let combinedScorePeriod = loadDash.cloneDeep(scorePeriodA);
-    combinedScorePeriod.studentScores = new Map<string, StudentScoreData>();
-
-    const periodAKeys = [...scorePeriodA.studentScores.keys()]
-    const periodBKeys = [...scorePeriodB.studentScores.keys()]
-
-    periodAKeys.forEach(key => {
-        combinedScorePeriod.studentScores.set(key, EMPTY_SCORE_VALUE)
-    });
-
-    periodBKeys.forEach(key => {
-        combinedScorePeriod.studentScores.set(key, EMPTY_SCORE_VALUE)
-    });
-
-    const unionedKeys = [...combinedScorePeriod.studentScores.keys()]
-
-    unionedKeys.forEach((key) => {
-
-        const aScoreData = scorePeriodA.studentScores.get(key)
-        const bScoreData = scorePeriodB.studentScores.get(key)
-
-        if(aScoreData) {
-            combinedScorePeriod.studentScores.set(key, addStudentScores(combinedScorePeriod.studentScores.get(key) as StudentScoreData, aScoreData, combinedScorePeriod.maxPoints))
-        }
-        if(bScoreData) {
-            combinedScorePeriod.studentScores.set(key, addStudentScores(combinedScorePeriod.studentScores.get(key) as StudentScoreData, bScoreData, combinedScorePeriod.maxPoints))
-        }
-    })
-
-    return combinedScorePeriod;
-}
-
-function addStudentScores(studentScoreA: StudentScoreData, studentScoreB: StudentScoreData, periodMax: number): StudentScoreData {
-
-    return {
-        score: Math.min(studentScoreA.score + studentScoreB.score, periodMax),
-        numPosts: studentScoreA.numPosts + studentScoreB.numPosts,
-        numIncomPost: studentScoreA.numIncomPost + studentScoreB.numIncomPost,
-        numComments: studentScoreA.numComments + studentScoreB.numComments,
-        numIncomComment: studentScoreA.numIncomComment + studentScoreB.numIncomComment,
-        awardsRecieved: studentScoreA.awardsRecieved + studentScoreB.awardsRecieved,
-        penaltiesRecieved: studentScoreA.penaltiesRecieved + studentScoreB.penaltiesRecieved,
     }
 }
